+++ server/src/api/modules/competitions/competition.router.ts
@@ -3,8 +3,7 @@ import { z } from 'zod';
 import { CompetitionCSVTableType, CompetitionStatus, CompetitionType, Metric, Team } from '../../../utils';
 import { checkAdminPermission, checkCompetitionVerificationCode } from '../../util/middlewares';
 import { executeRequest, validateRequest } from '../../util/routing';
-import { getDateSchema, teamSchema } from '../../util/schemas';
-import { getPaginationSchema } from '../../util/validation';
+import { getPaginationSchema, getDateSchema, teamSchema } from '../../util/validation';
 import { addParticipants } from './services/AddParticipantsService';
 import { addTeams } from './services/AddTeamsService';
 import { createCompetition } from './services/CreateCompetitionService';
