+++ server/src/api/modules/groups/group.router.ts
@@ -3,8 +3,7 @@ import { z } from 'zod';
 import { GroupRole, Metric, MigrationDataSource, Period } from '../../../utils';
 import { checkAdminPermission, checkGroupVerificationCode } from '../../util/middlewares';
 import { executeRequest, validateRequest } from '../../util/routing';
-import { getDateSchema, memberSchema, socialLinksSchema } from '../../util/schemas';
-import { getPaginationSchema } from '../../util/validation';
+import { getPaginationSchema, getDateSchema, memberSchema, socialLinksSchema } from '../../util/validation';
 import { findGroupCompetitions } from '../competitions/services/FindGroupCompetitionsService';
 import { addMembers } from './services/AddMembersService';
 import { changeMemberRole } from './services/ChangeMemberRoleService';
