+++ server/src/api/modules/players/player.router.ts
@@ -5,8 +5,7 @@ import { CompetitionStatus, Metric, Period } from '../../../utils';
 import { NotFoundError, ServerError } from '../../errors';
 import { checkAdminPermission, detectRuneLiteNameChange } from '../../util/middlewares';
 import { executeRequest, validateRequest } from '../../util/routing';
-import { getDateSchema } from '../../util/schemas';
-import { getPaginationSchema } from '../../util/validation';
+import { getPaginationSchema, getDateSchema } from '../../util/validation';
 import { findPlayerAchievementProgress } from '../achievements/services/FindPlayerAchievementProgressService';
 import { findPlayerAchievements } from '../achievements/services/FindPlayerAchievementsService';
 import { findPlayerParticipations } from '../competitions/services/FindPlayerParticipationsService';
